# swagger/swagger-output.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My API",
        "description": "Description",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:2222/"
        }
    ],
    "paths": {
...
        "/api/facilities/": {
            "get": {
                "description": "",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "List of all facilities",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "serverTime": {
                                            "type": "string",
                                            "example": "1"
                                        },
                                        "data": {
                                            "type": "object",
                                            "properties": {
                                                "facilities": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/Facility"
                                                    }
                                                }
                                            },
                                            "required": ["facilities"]
                                        }
                                    },
                                    "required": [
                                        "status",
                                        "serverTime",
                                        "data"
                                    ],
                                    "xml": {
                                        "name": "main"
                                    }
                                }
                            },
                            "application/xml": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "serverTime": {
                                            "type": "string",
                                            "example": "1"
                                        },
                                        "data": {
                                            "type": "object",
                                            "properties": {
                                                "facilities": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/Facility"
                                                    }
                                                }
                                            },
                                            "required": ["facilities"]
                                        }
                                    },
                                    "required": [
                                        "status",
                                        "serverTime",
                                        "data"
                                    ],
                                    "xml": {
                                        "name": "main"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "post": {
                "description": "",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "Facility successfully added",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "serverTime": {
                                            "type": "string",
                                            "example": "1"
                                        },
                                        "data": {
                                            "type": "object",
                                            "properties": {
                                                "facility": {
                                                    "$ref": "#/components/schemas/Facility"
                                                }
                                            },
                                            "required": ["facility"]
                                        }
                                    },
                                    "required": [
                                        "status",
                                        "serverTime",
                                        "data"
                                    ],
                                    "xml": {
                                        "name": "main"
                                    }
                                }
                            },
                            "application/xml": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "serverTime": {
                                            "type": "string",
                                            "example": "1"
                                        },
                                        "data": {
                                            "type": "object",
                                            "properties": {
                                                "facility": {
                                                    "$ref": "#/components/schemas/Facility"
                                                }
                                            },
                                            "required": ["facility"]
                                        }
                                    },
                                    "required": [
                                        "status",
                                        "serverTime",
                                        "data"
                                    ],
                                    "xml": {
                                        "name": "main"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "name": {
                                        "example": "any"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/facilities/{facilityId}/occupancies": {
            "get": {
                "description": "",
                "parameters": [
                    {
                        "name": "facilityId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Facility occupied status",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "serverTime": {
                                            "type": "string",
                                            "example": "1"
                                        },
                                        "data": {
                                            "type": "object",
                                            "properties": {
                                                "isOccupied": {
                                                    "type": "string",
                                                    "example": "true"
                                                },
                                                "occupancies": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/FacilityOccupancy"
                                                    }
                                                }
                                            },
                                            "required": [
                                                "isOccupied",
                                                "occupancies"
                                            ]
                                        }
                                    },
                                    "required": [
                                        "status",
                                        "serverTime",
                                        "data"
                                    ],
                                    "xml": {
                                        "name": "main"
                                    }
                                }
                            },
                            "application/xml": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "serverTime": {
                                            "type": "string",
                                            "example": "1"
                                        },
                                        "data": {
                                            "type": "object",
                                            "properties": {
                                                "isOccupied": {
                                                    "type": "string",
                                                    "example": "true"
                                                },
                                                "occupancies": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/FacilityOccupancy"
                                                    }
                                                }
                                            },
                                            "required": [
                                                "isOccupied",
                                                "occupancies"
                                            ]
                                        }
                                    },
                                    "required": [
                                        "status",
                                        "serverTime",
                                        "data"
                                    ],
                                    "xml": {
                                        "name": "main"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/facilities/release/{id}": {
            "post": {
                "description": "",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Facility released",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "serverTime": {
                                            "type": "string",
                                            "example": "1"
                                        },
                                        "data": {
                                            "type": "object",
                                            "properties": {
                                                "isOccupied": {
                                                    "type": "string",
                                                    "example": "false"
                                                }
                                            },
                                            "required": ["isOccupied"]
                                        }
                                    },
                                    "required": [
                                        "status",
                                        "serverTime",
                                        "data"
                                    ],
                                    "xml": {
                                        "name": "main"
                                    }
                                }
                            },
                            "application/xml": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "serverTime": {
                                            "type": "string",
                                            "example": "1"
                                        },
                                        "data": {
                                            "type": "object",
                                            "properties": {
                                                "isOccupied": {
                                                    "type": "string",
                                                    "example": "false"
                                                }
                                            },
                                            "required": ["isOccupied"]
                                        }
                                    },
                                    "required": [
                                        "status",
                                        "serverTime",
                                        "data"
                                    ],
                                    "xml": {
                                        "name": "main"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Facility": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "example": "1"
                    },
                    "name": {
                        "type": "string",
                        "example": "Facility name"
                    },
                    "occupancy": {
                        "type": "array",
                        "items": {
                            "xml": {
                                "name": "facilityoccupancy"
                            },
                            "$ref": "#/components/schemas/FacilityOccupancy"
                        }
                    }
                },
                "required": ["id", "name", "occupancy"],
                "xml": {
                    "name": "Facility"
                }
            },
...
            "FacilityOccupancy": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "example": "1"
                    },
                    "isOccupied": {
                        "type": "boolean",
                        "example": true
                    },
                    "facilityId": {
                        "type": "string",
                        "example": "1"
                    },
                    "facility": {
                        "xml": {
                            "name": "facility"
                        },
                        "$ref": "#/components/schemas/Facility"
                    },
                    "createdAt": {
                        "type": "number",
                        "example": 1577836800000
                    },
                    "relatedEntryId": {
                        "type": "string",
                        "example": "2"
                    }
                },
                "required": [
                    "id",
                    "isOccupied",
                    "facilityId",
                    "facility",
                    "createdAt",
                    "relatedEntryId"
                ],
                "xml": {
                    "name": "FacilityOccupancy"
                }
            },
...
        }
    }
}